Py3, restructure, GCM Encryption
Quite large restructuring, all parts of the inform parsing is moved to
unifi_protocol, so no protocol handling is done in unifi_gateway
also made it py3 compatible and added the missing GCM encryption method.

There are still some code missing to make it deliver the correct data
to the controller.